feat: CRUD işlemleri tamamlandı, update ve delete metodları service ve controller katmanlarına eklendi

+++ src/main/java/com/project/stokyonetim/service/CustomerService.java
@@ -5,28 +5,42 @@ import com.project.stokyonetim.repository.CustomerRepository;
 import org.springframework.stereotype.Service;
 import java.util.List;
 
-// Bu sinifin is mantigini yürüten bir servis katmani oldugunu Spring Boot'a bildirir
 @Service
 public class CustomerService {
 
-    // Veri tabani sorgularini yapabilmek icin repository katmanini iceriye bagliyoruz (Dependency Injection)
     private final CustomerRepository customerRepository;
 
     public CustomerService(CustomerRepository customerRepository) {
         this.customerRepository = customerRepository;
     }
 
-    // Postman'den gelen yeni musteri verisini veri tabanina kaydeder
     public Customer saveCustomer(Customer customer) {
         return customerRepository.save(customer);
     }
 
-    // Veri tabanindaki tum musterileri liste halinde Postman'e doner
     public List<Customer> getAllCustomers() {
         return customerRepository.findAll();
     }
+
+    // GÜVENLİ SİLME METODU
     public void deleteById(Long id) {
+        if (!customerRepository.existsById(id)) {
+            throw new RuntimeException("Silinecek müşteri bulunamadı! ID: " + id);
+        }
         customerRepository.deleteById(id);
     }
 
+    // GÜNCELLEME METODU
+    public Customer updateCustomer(Long id, Customer customerDetails) {
+        Customer existingCustomer = customerRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Müşteri bulunamadı! ID: " + id));
+
+        existingCustomer.setFirstName(customerDetails.getFirstName());
+        existingCustomer.setLastName(customerDetails.getLastName());
+        existingCustomer.setEmail(customerDetails.getEmail());
+        existingCustomer.setPhoneNumber(customerDetails.getPhoneNumber());
+        existingCustomer.setGender(customerDetails.getGender());
+
+        return customerRepository.save(existingCustomer);
+    }
 }
